pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Victo Almeida <victorres.almeida@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
Django = "^4.0.5"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
